requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
pymongo==4.6.0
motor==3.3.2
neo4j==5.14.1
redis==5.0.1

# NLP & ML
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2
ollama==0.1.6
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.18

# Social Media APIs
praw==7.7.1
python-telegram-bot==20.7
feedparser==6.0.10

# Web Scraping (instead of Twitter API)
playwright==1.40.0
newspaper3k==0.2.8
nest-asyncio==1.5.8

# WhatsApp Integration
twilio==8.10.3

# Image/Video Processing
pillow==10.1.0
opencv-python==4.8.1.78
imagehash==4.3.1

# Fact-checking & Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
aiohttp==3.9.1
httpx==0.25.2
celery==5.3.4
python-multipart==0.0.6

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
